third_party/WebKit/LayoutTests/http/tests/devtools/editor/text-editor-auto-whitespace-removing-expected.txt

CONSOLE WARNING: line 3078: Element.createShadowRoot is deprecated and will be removed in M73, around March 2019. Please use Element.attachShadow instead. See https://www.chromestatus.com/features/4507242028072960 for more details.
This test verifies that auto-appended spaces are removed on consequent enters.

function (){}
    if (a == b) {


Running: testCollapsedBlock
function.(){

....|
}
....if.(a.==.b).{


Running: testOpenCurlyBrace
function.(){}
....if.(a.==.b).{

........|


Running: testSmartIndent
function.(){}
..

..|..if.(a.==.b).{


Running: testMultiCursorSelection
function.(){}
..

..|..

....|if.(a.==.b).{


Running: testEditedAutoIndent
function.(){}
....if.(a.==.b).{
........W
........|
